Refactpred role service tests into different class
